test: add tests for getHeadersConfig
