+++ lib/parseOptions.js
@@ -4,6 +4,7 @@ import { resolve } from 'path'
 import moment from 'moment'
 
 import { version } from '../package'
+import { getHeadersConfig } from './utils/headers'
 import { proxyStringToObject, agentFromProxy } from 'contentful-batch-libs/dist/proxy'
 
 const SUPPORTED_ENTITY_TYPES = [
@@ -15,47 +16,6 @@ const SUPPORTED_ENTITY_TYPES = [
   'webhooks',
   'editorInterfaces'
 ]
-
-/**
- * Turn header option into an object. Invalid header values
- * are ignored.
- *
- * @example
- * getHeadersConfig('Accept: Any')
- * // -> {Accept: 'Any'}
- *
- * @example
- * getHeadersConfig(['Accept: Any', 'X-Version: 1'])
- * // -> {Accept: 'Any', 'X-Version': '1'}
- *
- * @param value {string|string[]}
- */
-function getHeadersConfig (value) {
-  if (!value) {
-    return {}
-  }
-
-  const values = Array.isArray(value) ? value : [value]
-
-  return values.reduce((headers, value) => {
-    value = value.trim()
-
-    const separatorIndex = value.indexOf(':')
-
-    // Invalid header format
-    if (separatorIndex === -1) {
-      return headers
-    }
-
-    const headerKey = value.slice(0, separatorIndex).trim()
-    const headerValue = value.slice(separatorIndex + 1).trim()
-
-    return {
-      ...headers,
-      [headerKey]: headerValue
-    }
-  }, {})
-}
 
 export default function parseOptions (params) {
   const defaultOptions = {
